KAFKA-12637: Remove deprecated PartitionAssignor interface
